Refactor career components
Break down CareerPostList and CareerPostEditor into smaller, reusable components.
