+++ src/components/admin/careers/components/JobStatus.tsx
@@ -1,0 +1,58 @@
+
+import React from 'react';
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Card, CardContent } from "@/components/ui/card";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { CareerPost } from '../types';
+
+interface JobStatusProps {
+  currentPost: CareerPost;
+  setCurrentPost: React.Dispatch<React.SetStateAction<CareerPost | null>>;
+}
+
+const JobStatus: React.FC<JobStatusProps> = ({ currentPost, setCurrentPost }) => {
+  return (
+    <Card className="border border-gray-200">
+      <CardContent className="p-6">
+        <h4 className="font-medium text-[#228977] mb-4">Listing Status</h4>
+        
+        <div className="space-y-4">
+          <div>
+            <Label htmlFor="status" className="text-sm font-medium text-[#21665a]">
+              Status
+            </Label>
+            <Select 
+              value={currentPost.status}
+              onValueChange={(value: 'active' | 'draft' | 'closed') => setCurrentPost({...currentPost, status: value})}
+            >
+              <SelectTrigger className="mt-1.5 border-gray-300">
+                <SelectValue placeholder="Select status" />
+              </SelectTrigger>
+              <SelectContent>
+                <SelectItem value="active">Active</SelectItem>
+                <SelectItem value="draft">Draft</SelectItem>
+                <SelectItem value="closed">Closed</SelectItem>
+              </SelectContent>
+            </Select>
+          </div>
+          
+          <div>
+            <Label htmlFor="postDate" className="text-sm font-medium text-[#21665a]">
+              Publish Date
+            </Label>
+            <Input 
+              id="postDate"
+              type="date"
+              value={currentPost.date}
+              onChange={(e) => setCurrentPost({...currentPost, date: e.target.value})}
+              className="mt-1.5 border-gray-300"
+            />
+          </div>
+        </div>
+      </CardContent>
+    </Card>
+  );
+};
+
+export default JobStatus;
